refactor search components
