+++ packages/ndla-ui/src/Search/component.search.scss
@@ -3,3 +3,5 @@
 @import './component.search-overlay';
 @import './component.active-filters';
 @import './component.search-page';
+@import './component.search-result';
+@import './component.search-filter';
